backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-cors-headers
